Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/sanselan/trunk@1308079 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/sanselan/formats/tiff/TiffField.java
@@ -290,7 +290,7 @@ public class TiffField implements TiffConstants
         // Debug.debug("fillInValue valueLength", valueLength);
 
         if (valueOffset < 0 ||
-            ((long)valueOffset) + valueLengthLong > byteSource.getLength()) {
+            valueOffset + valueLengthLong > byteSource.getLength()) {
             throw new TiffValueOutsideFileBoundsException(
                 "Attempt to read byte range starting from " + valueOffset + " " +
                 "of length " + valueLength + " " +
